client/requirements.txt
anyio==3.5.0
asyncclick==8.0.3.2
idna==3.3
importlib-metadata==4.11.3
python-kasa==0.4.3
sniffio==1.2.0
zipp==3.8.0
